+++ bubo/help_strings.py
@@ -14,20 +14,16 @@ More help on commands or subcommands using 'help' as the next parameter.
 For source code, see https://github.com/elokapina/bubo
 """
 
-HELP_GROUPJOIN = """Join one or more users to a predefined group of rooms.
+HELP_GROUPINVITE = """Invite a user to a predefined group of rooms.
 
 Syntax:
 
-    groupjoin groupname @user1:domain.tld @user2:domain.tld
+    groupinvite @user1:domain.tld groupname [subgroups]
 
-Where "groupname" should be replaced with the group to join to. Groups can be listed
-by giving the command without any extra parameters, eg just "groupjoin".
+Where "groupname" should be replaced with the group to join to. Additionally
+subgroups can be given as well.
 
 Groups must be configured to the Bubo configuration file by an administrator.
-
-If Bubo has Synapse admin powers, it will try to join admin API join any local users
-(which still requires Bubo to be in the room and be able to invite).
-Otherwise, a normal onvitation is used.
 
 This command requires coordinator level permissions.
 """