+++ CHANGELOG.md
@@ -18,9 +18,8 @@
 * Add `join` command, which allows (Synapse admin) force joining users to rooms or as 
   a fallback inviting users to rooms.
 
-* Add `groupjoin` command, which allows (Synapse admin) force joining users to a group
-  of rooms or as a fallback inviting users to rooms. Room groups are preconfigured in
-  the config file.
+* Add `groupinvite` command, which allows inviting a user to a group
+  of rooms. Room groups are preconfigured in the config file.
 
 ### Changed